+++ webpack/webpack.analyzer.ts
@@ -1,11 +1,27 @@
-import webpack from 'webpack';
-import merge from 'webpack-merge';
-import common from './webpack.common';
 import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer';
+import TsconfigPathsPlugin from 'tsconfig-paths-webpack-plugin';
 
-const analyzer: webpack.Configuration = merge(common, {
+const analyzer = {
   mode: 'production',
+  module: {
+    rules: [
+      {
+        exclude: /node_modules/,
+        test: /\.ts(x?)$/,
+        use: [{ loader: 'ts-loader' }],
+      },
+      {
+        enforce: 'pre',
+        loader: 'source-map-loader',
+        test: /\.js$/,
+      },
+    ],
+  },
   plugins: [new BundleAnalyzerPlugin()],
-});
+  resolve: {
+    extensions: ['.ts', '.tsx', '.js'],
+    plugins: [new TsconfigPathsPlugin({})],
+  },
+};
 
 export default analyzer;