+++ webpack/webpack.prod.ts
@@ -1,12 +1,24 @@
-import webpack from 'webpack';
-import merge from 'webpack-merge';
-import common from './webpack.common';
+import TsconfigPathsPlugin from 'tsconfig-paths-webpack-plugin';
 import path from 'path';
 
 const __dirname = path.resolve();
 
-const prod: webpack.Configuration = merge(common, {
+const prod = {
   mode: 'production',
+  module: {
+    rules: [
+      {
+        exclude: /node_modules/,
+        test: /\.ts(x?)$/,
+        use: [{ loader: 'ts-loader' }],
+      },
+      {
+        enforce: 'pre',
+        loader: 'source-map-loader',
+        test: /\.js$/,
+      },
+    ],
+  },
   optimization: {
     chunkIds: "named",
     splitChunks: {
@@ -32,6 +44,10 @@ const prod: webpack.Configuration = merge(common, {
     path: path.join(__dirname, "dist/prod"),
     publicPath: '/',
   },
-});
+  resolve: {
+    extensions: ['.ts', '.tsx', '.js'],
+    plugins: [new TsconfigPathsPlugin({})],
+  },
+};
 
 export default prod;